[FE] refactor: cameraSetting 분리
